resolve all StyleCop.CSharp complains
